moved components into more sensible locations
