+++ src/apps/hasabe/App.tsx
@@ -1,6 +1,5 @@
-import { Theme } from "components";
+import { AuthProvider, AuthWrapper, Login, Theme } from "components";
 
-import { AuthProvider, AuthWrapper, Login } from "./components/Auth";
 import { TaskUi } from "./components/Tasks";
 
 const App = () => {
