+++ src/components/index.ts
@@ -1,5 +1,6 @@
-export { default as Theme } from "./Theme";
 export { default as CardButton } from "./CardButton";
-export { default as UtilityBar } from "./UtilityBar";
-export { default as KeyboardShortcut } from "./KeyboardShortcut";
 export { default as ContextMenuItem } from "./ContextMenuItem";
+export { default as KeyboardShortcut } from "./KeyboardShortcut";
+export { default as Theme } from "./Theme";
+export { default as UtilityBar } from "./UtilityBar";
+export * from "./Auth";
